+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.09.29',
+  '2022.10.06',
   // Fallback runtime versions.
+  '2022.09.29',
   '2022.09.26',
-  '2022.09.14',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -55,7 +55,7 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.18.0';
 final String toolStableDartSdkVersion = '2.18.2';
-final String toolStableFlutterSdkVersion = '3.3.3';
+final String toolStableFlutterSdkVersion = '3.3.4';
 final String toolPreviewDartSdkVersion = '2.19.0-146.2.beta';
 final String toolPreviewFlutterSdkVersion = '3.4.0-17.2.pre';