+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.10.06`.
+ * Upgraded stable Flutter analysis SDK to `3.3.4`.
 
 ## `20221004t093700-all`